10.project/11.mathqna_app/requirements.txt
Flask==2.3.3
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.3.7
wikipedia==1.4.0
arxiv==1.4.8
numexpr==2.8.7
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
